screw up between int and byte
git-svn-id: http://svn.vuze.com/public/client@15375 c2c1b7af-1477-405b-86ee-8aefd1fc4dd2

+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/control/impl/DHTControlImpl.java
@@ -1094,7 +1094,7 @@ DHTControlImpl
 					return( basis.getLifeTimeHours());
 				}
 				
-				public int
+				public byte
 				getReplicationFactor()
 				{
 					return( basis.getReplicationFactor());
@@ -4228,7 +4228,7 @@ DHTControlImpl
 			return( delegate.getLifeTimeHours());
 		}
 		
-		public int 
+		public byte 
 		getReplicationFactor() 
 		{
 			return( delegate.getReplicationFactor());
