+++ trunk/azureus2/src/com/aelitis/azureus/core/dht/transport/DHTTransportValue.java
@@ -51,7 +51,7 @@ DHTTransportValue
 	public int
 	getLifeTimeHours();
 	
-	public int
+	public byte
 	getReplicationFactor();
 
 	public String
